[CHORE] ArtistScreenViewModel -> ArtistViewModel 이름 변경
